Update mongodb_test
Corrected pymongo import issue in Visual Studio (thanks to an awesome co-worker's help and expertise). Issue was VS was using the path usr/bin/python3 but pymongo was not imported there.  Pymongo was actucally loaded to path usr/local/python3 in VS, and when manually changing VS to use that path, the file ran successfully. Co-worker corrected the file path in VS to import pymongo to usr/bin/python3 (which took about 30 minutes to debug). Pymongo is now imported in usr/bin/python3 and the file successfully ran creating the collection "students".